Remove new Paper configuration and workaround for SkinsRestorer (#177)
Fixes #169

Basically this class was added for ViaVersion to automatically disable its packet limiter because we have a built-in packet limiter. However, SkinsRestorer and EssentialsX apparently use this class to check the server version, if they detect that the class exists, they try to enable features from newer versions (which PandaSpigot does not have), which results in errors. Since the class was removed, the workaround for SkinsRestorer became unnecessary. (I tested it and it worked perfectly)

Warning: Removing this class is unlikely to cause problems with ViaVersion, however, if it does, you can disable the ViaVersion packet limiter in the plugin configuration.