Use downloads API for version checking
Instead of using the GitHub Actions API, now use the downloads.hpfxd.com
bibliothek instance to check how far behind the current version is from the
latest build. This is the case with modern versions of Paper.

Additionally, change the README to link to a direct download for the latest build